Create newRecon
Applescript (saves as application within script editor when in use)
Provides a graphical interface for setting specific extension attributes
This was created to allow machine type (a Jamf Extension Attribute set in the JSS with a pull down menu) to be set to Classroom/Lab, Faculty, Staff, Research, Loaner, etc. 
These machine types are used to scope software that may only be licensed for academic use appropriately.